Add method for logging in a user with a test request
